[MOBILESDK-170] Android - Implement AIMS Web Login
I've added UI sample for PKCE auth. I've done a very simple MVVM architecture if you can call it that. Basically I have some sort of god viewmodel that handles all the necessary actions regarding the PKCE auth. I've choose this path just to keep everything simple enough to read and understand. After all, this is a showcase app.

This first version of the sample app contains the alfresco core and auth modules that are contained on maven local and need to be generated on every device this code should run on.