[MOBILESDK-222] - Sample app is crashing when pressing on Login after visiting advanced settings screen
Apparently there's some sort of an issue when trying to navigate with an already defined action while coming back from another activity. Just to elaborate, we have the main activity inside the sample app, but when trying to auth the user, we're starting another activity that contains the webview (this is inside the auth sdk). When coming back from that activity the navigation component doesn't seem to find the defined navigation action. As a fast workaround for this, I'm using deeplinks instead of statically defined navigation actions.